Implementing the vector field Niagara data interface, for both CPU and GPU particles.
Cascade only supported GPU vector fields, which means that the bulk vector field data stored in UVectorFieldStatic::SourceData isn't intended to be readable at runtime. As a result I have added a boolean flag bAllowCPUAccess to the UVectorFieldStatic class, and if this flag is set it makes a copy of the bulk data and stores it in UVectorFieldStatic::CPUData. This data is converted from the 16-bit floats point source data to 32-bit floats. This array is marked as transient and won't be serialized. The copied data is only available to use if bAllowCPUAccess is true. If bAllowCPUAccess is false it will clear the CPUData array. This copied data is used in UNiagaraDataInterfaceVectorField::SampleVectorField when sampling on the CPU.

The UNiagaraDataInterfaceVectorField class is overhauled. There are no longer local copies of the dependent vector field properties from the UVectorField pointer. Instead these quantities are accessed directly through the pointer if it is non-null in the relevant functions. The InitField method is therefore removed, as it now redundant.

GPU particles sample the 3D texture provided by the static vector field render resource, so there's no need for an additional copy per DI.

The old sampling function is completely rewritten to be consistent and reach parity with the built-in GPU texture sampler.

To make sure the render thread is synced to the game thread state there is a FlushRenderingCommands call in UNiagaraDataInterfaceVectorField::PreEditChange. A better long term solution would be to use something like a RT proxy.

The Lock and Unlock methods are not needed anymore since we now use the copied bulk data instead of (incorrectly) accessing the bulk data.

A simple GetErrors() method is implemented, to show an error if bAllowCPUAccess isn't set and if there's invalid vector field data.

All deprecated code and commented out blocks of old code are removed for clarity.

Updates GRegisterNiagaraCustomVersion to make sure any old script is forced to recompiled due to serialization changes.

#rb shaun.kime simon.tovey

[CL 4550841 by Morten Vassvik in Dev-Niagara branch]