Implemented a new and improved Curl Noise Niagara data interface.
The main motivation behind curl noise is that "the curl of any continuous vector field is divergence free". Incidentally it also results in a field containing vortices of approximately even size, which is nice for turbulence approximations.

The old implementation used a hard-coded 15x15x15 volume texture to store the curl noise, which in turn was calculated from a coarsely sampled gradient noise field.

This approach have several problems:
 - The resulting noise is not divergence-free, both due to the finite resolution of the backing noise field and the coarse finite difference used to calculate the curl.
 - The hard-coded (and low) resolution field limits the size of vortices/coils, and is easily visible while tiling the field.
 - You need high enough resolution to capture the size of the vortices and also keep the divergence low.
 - Requires recomputation to re-seed.
 - Takes up memory.

But, it's also extremely cheap to evaluate at runtime, as it's essentially one trilinear texture sampling.

This new approach uses procedural noise to generate curl noise.

3D simplex noise is used as the background noise field, whose derivatives are evaluated analytically (essentially for free). There are plenty of upsides, mirroring the problems of the texture approach:
 - There's no need for finite difference evaluations of derivatives, and is automatically divergence free at any point.
 - There's no limits on scale, large and small, the field is essentially infinite (except for floating point precision), removing the need for tiling.
 - Re-seeding has next to no cost and is simply another parameter that's added to the sample position.
 - There's no memory cost

On the other hand, it's more expensive to sample than a simple texture read. It's also trivially parallelizable as long as the RNG used for the noise isn't sequential in nature.

In the end it's a tradeoff of quality over (runtime) performance. When performance is truly necessary, and there's memory to spare, the user can simply bake the curl field into a texture and read it as a vector field instead.

The CPU-side simplex noise code itself is a direct port of the equivalent HLSL code in Engine/Shaders/Private/Random.ush and is identical in functionality and output except for minor floating point precision differences. The CPU-side implementation could likely be optimized further, as there are difference considerations on the CPU than the GPU.

Caveat: There's a risk of getting floating point precision artifacts if the input positions gets too large.

#rb shaun.kime, simon.tovey

[CL 4568639 by Morten Vassvik in Dev-Niagara branch]