Adds support for deterministic random number generation for both CPUs and GPUs.
Both backend changes and assets are included.


**Frontend stuff**:

Determinism is controllable on a per-emitter basis using the "Determinism" boolean toggle found in Emitter Properties in the Emitter Update stack. This defaults to *OFF*.

The "Random Seed" of the deterministic RNG while Determinism is can be controlled in Emitter Properties, and is editable while Determinism is toggled.

The Niagara Function script "Random Range" is reworked to support both non-deterministic and deterministic uniform random number generation. This is the primary and recommended way for the user to generate random numbers, and all other functions, dynamic inputs or modules that generate random numbers should call this function either explicitly or through some other function or dynamic input which calls it. The function takes a minimum and maximum Numeric value as input (as before), but now also expose extra *hidden* inputs to optionally control the random number generation on a per-module basis:
	- Randomness Mode: An Enum, whose entries are
		- Simulation Default: Use the per-emitter settings. This is the default value.
		- Deterministic: Override using the deterministic RNG
		- Non-deterministic: Override using the non-deterministic RNG
	- Seed: A seed to provide to the deterministic RNG. Ignored if determinism is disabled.
	- Override Seed: Whether to let the deterministic RNG use the provided seed. This is *OFF* by default.

The provided selection of dynamic input scripts (in the Uniform Range folder) and modules are updated to use Random Range. The randomness can be controlled through the new "Randomness" category of module input variables included for all of these, enabling the user to provide fine control of the RNG behavior.

Random Range calls a new "Seeded Random" operator, which calls to the appropriate VM code based on the inferred input types. It has 3 integer inputs that can be used to explicitly seed the RNG instead of relying on Random Range passing the Particle ID and the Emitter/System tick count as seeds explicitly.


**Significant changes to assets:**
 - GaussianRandomFloat is rewritten to be more optimized. Produces two Gaussian numbers at the same time instead of just one.
 - RandomQuaternion: An error is fixed, and the code is rewritten to be more efficient.
 - RandomUnitVector: A replacement for RandomVector with a more self-describing name.
 - BoxLocation: Rewritten to correctly distribute particles uniformly in a wall of a certain thickness on demand. Can be controlled by the Thickness Mode module variable.
 - SphereLocation: Rewritten to correctly distribute particles uniformly in a spherical shell of a certain thickness on demand. Can be controlled by the Thickness Mode module variable.
 - CylinderLocation: Rewritten so that the wall thickness can be controlled by the Thickness Mode module variable.

In general all the assets related to random number generation should be changed so that the determinism can be overridden.

**New assets:**
 - RandomParticleInBox: Helper for the BoxLocation module.
 - RandomParticleInSphere: Helper for the SphereLocation module.
 - ENiagaraRandomnessMode: Enum enum used to select determinism mode for overwriting the per-emitter settings.
 - ENiagaraExpansionMode: Controlls the Thickness expansion mode for the Location

**Deprecated assets**:
 - LinearColorAlongVector (Dynamic Input): Replaced by UniformRangedLinearColor
 - CurlNoiseLocation (Module): Replaced by SampleCurlNoise
 - RandomVector (function): Replaced by RandomUnitVector

**Tests**:
 - RandomnessCPUAndGPUTest: This test calls Seeded Random directly and tests int32, Vector, Vector2D and Vector4 calls to Seeded Random, for both emitters and particles scripts. This is primarily to test the consistency of the VM code.


**Backend stuff**:
The random number generator is implemented in NiagaraEmitterInstanceShader.usf as overloaded functions to the "rand" call. The deterministic versions take 4 parameters, one for scaling and three seeds. The non-deterministic functions take a single parameter, a scaling parameter.

The deterministic RNG is based on NiagaraRand4DPCG32 and NiagaraRand3DPCG32 from Random.ush, but are modified slightly to cover our needs.  It uses only integer operations (except for the float normalization on return) and should be consistent across platforms and devices.

This deterministic RNG is slightly more costly than the old CPU for a single component, but should be cheaper on multiple components as long as the operations are vectorized properly. Likewise there should be a lot of opportunity for constant folding that will offload some of the work to compile time optimizations.

A few additional changes have been made:
 - Engine.Emitter.TotalSpawnedParticles is added to count the number of spawned particles since the simulation started. This is re-initialized upon looping or reset.
 - Engine.System.TickCount is added to be used as an implicit seed for the deterministic RNG.
 - Emitter.Determinism is added to the emitter properties. This can be used to controll the RNG behavior. Implemented similar to Emitter.Localspace.
 - Emitter.RandomSeed is added to the emitter properties. This can be used to seed the deterministic RNG, unless otherwise overridden.
 - Particles.UniqueID is added as a engine-controlled attribute. Calculated as Particles.UniqueID = Engine.Emitter.TotalSpawnedParticles + ExecIndex() in the particle spawn scripts.

Particles.UniqueID, Engine.System.TickCount and Emitter.RandomSeed serve as the default seeds to the deterministic RNG.

Particles.UniqueID is different from Particles.ID as the use case is slightly different. First, it's only a single integer. Second, it's contiguous (monotonically increasing) and unique, and does not reuse indices unless it wraps around (4 billion particles, at which point there will not be any collisions anyway). This does not rely on the persistent IDs in any way and are calculated all the time regardless of whether they're used or not.

The custom script version GUID is updated to force a script recompile.


**Profiling**:
Preliminary testing shows that the deterministic RNG is about as costly as the non-deterministic RNG.


#tests Win64 cooked and uncooked automated tests pass locally.
#rb Shaun.Kime

[CL 4708835 by Morten Vassvik in Dev-Niagara branch]